+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 
 - gRPC client support
 - Added support to set queue options globally or per job [#158]
+- Add PSR-3 compatible logging with RoadRunner integration
 
 ## Unreleased